refactor: Simplify asset information handling and event bus integration in AiTab component
